requirements.txt
asgiref==3.2.7
Django==3.0.5
django-multiselectfield==0.1.12
pytz==2019.3
sqlparse==0.3.1
